Fix incorrect package skip when review failed
Previous to this commit, due to the fact how iteration over lists works
in Python, when a package failed a review it was directly removed from
the list of packages. This, however, led to the package list shortening
by 1 and the index of the next package for review would be lowered by
one. In the next iteration, the package with the next-higher index would
be reviewed, skipping the previously next package that now was shifted
to the index of the package dropped, leading to a package being built
without being reviewed. This fixes this behavior.