Support provides and unsatisfyable deps via pyalpm
This commit ports most of the pacman module to pyalpm, because it is
needed to sanely get package-provides in a dependency graph if a
provides that is not a package is referenced as a dependency.

In addition to that, blinky now recognizes if dependencies are
unsatisfiable by installed packages, packages in sync-DBs and PKGBUILDs
in the AUR and drops those subgraphs of the dependency tree before even
attempting to get user interaction for them, as required for blinky's
promise of minimal hassle.