implement tree deduplication
This commit adds functionality to deduplicate the built dependency tree
should it have identical packages as separate tree nodes.
This can happen as the built-up of the dependency tree is concurrent and
the check if a package object is already being built and starting
another object build is not atomic.
Deduplicating the tree afterwards is the less clean version conceptually,
but easier to maintain code-wise in the long run as we can continue to
use the concurrency mechanisms built into python.