Refresh package databases before package removal
In previous versions of blinky, when removing makedeps after their
purpose was fullfilled, it could not be verified which makedeps were
actually installed for this purpose as an older snapshot of the local
package databases was loaded. This commit fixes this by refreshing the
loaded package databases prior to checking which makedepends actually
got installed to remove them afterwards.