Batch requests to AUR-API
Before this commit, each package would query the AUR-API upon
construction. This, however, produces a significant number of requests.
To reduce this number, Packages only query the AUR-API if they might be
an AUR-package but have not been supplied with AUR-data. This AUR-data
is instead batch-requested by packages for all of their dependencies
and handed down the package graph.