add rebuild-script for blinky
In case a new Python version comes in, all python packages must be
rebuilt. In the repos this is taken care by the package maintainers, for
the AUR this would be blinky's job. As blinky itself is written in
Python, this is a chicken-egg-problem. Therefore this commit introduces
a lightweigth shell-wrapper to perform this rebuild for blinky, to then
be able to use blinky for the remaining rebuilds.