Restructure dependency handling
Restructuring dependency handling was necessary, as few, but existing
PKGBUILD-files require their dependencies also as makedeps. Therefore we
cannot build dependencies and the package itself at the same time if
both are from AUR, but to ensure all PKGBUILDs will build correctly, we
have to ensure to build the dependencies first, install them and then
build the actual packages we want to build.

In addition to that, as we now do require multistage builds,
makedepends-cleanup was separated out to do that just once at the very
end instead of multiple times, keeping user interaction to a minimum.