print information about dependency tree for AUR-dependencies
This commit will print information for non-explicitly installed AUR
packages, so that the user gets a feel for where they originate and why
they are pulled in. Only one dependency chain is explicitly printed, if
an AUR-dependency is pulled in by two packages simultaneously, only one
chain is shown exemplarily, but diverging branches are denoted with
an "(among others)"-suffix to the package.

Closes #15