Wrap calls to alpm-DBs into a mutex
We leverage concurrency heavily in blinky to construct the package
dependency graph. As a consequence, every package is on its own to get
its installed-state and similar things from the local package databases.

However, it turns out, that these calls to the package databases are not
thread safe, leading to seemingly random segfaults.

This commit wraps all of the potential candidates to produce such a
segfault into a mutex-lock, which should resolve and in the future
prevent this problem.

This mutex-lock should not significantly impact blinky, as the calls to
the local package database are not time critical, as, with regards to
time, dependency-tree construction is dominated by network-related delays.