treat checkdeps as makedeps
Up to this point, blinky somehow ignored checkdepends (which apparently
doesn't pose a problem for a significant number of packages in the AUR).
This commit now also installs checkdepends. They are simply treated as
makedepends, because albeit they are used for different parts of the
PKGBUILD, they are not much different on a technical level, so treating
them as makedepends gives us all the handling we want for checkdepends,
as it is identical to makedepends.